+++ source/Engine/Sprite.h
@@ -34,20 +34,20 @@ namespace CS230 {
         void Update(double dt) override;
         void Load(const std::filesystem::path& sprite_file, GameObject* _given_object);
         void Draw(Math::TransformationMatrix display_matrix);
-        Math::ivec2 GetHotSpot(int index);
+        Math::ivec2 GetHotSpot(size_t index);
         Math::ivec2 GetFrameSize();
 
-        void PlayAnimation(int animation);
+        void PlayAnimation(size_t animation);
         bool AnimationEnded();
 
-        const int& CurrentAnimation() const { return current_animation; }
+        const size_t& CurrentAnimation() const { return current_animation; }
     private:
-        Math::ivec2 GetFrameTexel(int index) const;
+        Math::ivec2 GetFrameTexel(size_t index) const;
         // Texture* texture;
         std::shared_ptr<CS230::Texture> texture;
         std::vector<Math::ivec2> hotspots;
 
-        int current_animation;
+        size_t current_animation;
         Math::ivec2 frame_size;
         std::vector<Math::ivec2> frame_texels;
         std::vector<Animation*> animations;
